requirements.txt
pytz
supabase
langchain_core
ta-lib
python-binance
langchain_openai
langchain_google_genai
python-dotenv
pytrends